+++ Game/src/game/level/Level.java
@@ -1,6 +1,6 @@
 package game.level;
 
-import game.Vector2i;
+import game.Vector2d;
 import game.entity.Player;
 import game.level.block.Block;
 
@@ -13,7 +13,7 @@ import javax.imageio.ImageIO;
 
 public class Level {
 	
-	public Vector2i startPosition;
+	public Vector2d startPosition;
 	
 	private int width, height;
 	
@@ -29,36 +29,36 @@ public class Level {
 	public static Level[] levels  = new Level[100];
 	
 	static {
-		levels[0] = new Level("/levels/level0.png", "/levels/overlay0.png", 0, new Vector2i(64*2, 64*5+32));
-		levels[1] = new Level("/levels/level1.png", "/levels/overlay1.png", 1, new Vector2i(64, 64*6+32));
-		levels[2] = new Level("/levels/level2.png", "/levels/overlay2.png", 2, new Vector2i(64, 64*6+32));
-		levels[3] = new Level("/levels/level3.png", 3, new Vector2i(64, 64*6+32));
-		levels[4] = new Level("/levels/level4.png", 4, new Vector2i(64, 64+32));
-		levels[5] = new Level("/levels/level5.png", 5, new Vector2i(64*2, 64*6+32));
-		levels[6] = new Level("/levels/level6.png", 6, new Vector2i(62*2, 64*5+32));
-		levels[7] = new Level("/levels/level7.png", 7, new Vector2i(64*1, 64*4+32));
-		levels[8] = new Level("/levels/level8.png", 8, new Vector2i(64*1, 64*3+32));
-        levels[9] = new Level("/levels/level9.png", 9, new Vector2i(64*13 + 32, 64*5 + 32));
-		levels[10] = new Level("/levels/level10.png", 10, new Vector2i(64*2, 64*7+32));
-		levels[11] = new Level("/levels/level11.png", 11, new Vector2i(64*2, 64*5+32));
-		levels[12] = new Level("/levels/level12.png", 12, new Vector2i(64*2, 64*6+32));
-		levels[13] = new Level("/levels/level13.png", 13, new Vector2i(64*1, 64*1+32));
-		levels[14] = new Level("/levels/level14.png", 14, new Vector2i(64*2, 64*4+32));
+		levels[0] = new Level("/levels/level0.png", "/levels/overlay0.png", 0, new Vector2d(64*2, 64*5+32));
+		levels[1] = new Level("/levels/level1.png", "/levels/overlay1.png", 1, new Vector2d(64, 64*6+32));
+		levels[2] = new Level("/levels/level2.png", "/levels/overlay2.png", 2, new Vector2d(64, 64*6+32));
+		levels[3] = new Level("/levels/level3.png", 3, new Vector2d(64, 64*6+32));
+		levels[4] = new Level("/levels/level4.png", 4, new Vector2d(64, 64+32));
+		levels[5] = new Level("/levels/level5.png", 5, new Vector2d(64*2, 64*6+32));
+		levels[6] = new Level("/levels/level6.png", 6, new Vector2d(62*2, 64*5+32));
+		levels[7] = new Level("/levels/level7.png", 7, new Vector2d(64*1, 64*4+32));
+		levels[8] = new Level("/levels/level8.png", 8, new Vector2d(64*1, 64*3+32));
+        levels[9] = new Level("/levels/level9.png", 9, new Vector2d(64*13 + 32, 64*5 + 32));
+		levels[10] = new Level("/levels/level10.png", 10, new Vector2d(64*2, 64*7+32));
+		levels[11] = new Level("/levels/level11.png", 11, new Vector2d(64*2, 64*5+32));
+		levels[12] = new Level("/levels/level12.png", 12, new Vector2d(64*2, 64*6+32));
+		levels[13] = new Level("/levels/level13.png", 13, new Vector2d(64*1, 64*1+32));
+		levels[14] = new Level("/levels/level14.png", 14, new Vector2d(64*2, 64*4+32));
 	}
 	
 	private static Level currentLevel = levels[0];
 	private static int currentLevelId = 0;
 	
-	public Level(String levelPath, int id, Vector2i startPosition){		
+	public Level(String levelPath, int id, Vector2d startPosition){
 		loadLevel(levelPath, id, startPosition);
 	}
 	
-	public Level(String levelPath, String overlayPath, int id, Vector2i startPosition){		
+	public Level(String levelPath, String overlayPath, int id, Vector2d startPosition){
 		loadOverlayImage(overlayPath);
 		loadLevel(levelPath, id, startPosition);
 	}
 	
-	private void loadLevel(String levelPath, int id, Vector2i startPosition){
+	private void loadLevel(String levelPath, int id, Vector2d startPosition){
 		this.id = id;
 		loadLevelImage(levelPath);
 		width = tilesImage.getWidth();
